disp: msm: sde: validate plane mode and gem obj flags
Add changes to validate the plane fb_translation mode of
drm_gem_obj attached to plane. It avoids device panic on S2
translation fault and fails the drm_atomic_commit for which
mismatch is detected. In current codeflow, only S1 mappings are
modified when dma_buf is detached from Non_sec CB and attached
to secure SB as part of msm_gem_get_iova_locked API, but S2
mapping entries are not modified and this causes crash.

Change-Id: Ib4c9fab4cd647d4ee3fb33f885ba2d578e2d46e7
Signed-off-by: Jayaprakash Madisetty <quic_jmadiset@quicinc.com>
Signed-off-by: Akash Gajjar <quic_agajjar@quicinc.com>