disp: msm: dp: clear link info capabilities during DP disconnect
This change clears all link info capabilities as a part of
panel deinitialisation during a disconnect sequence. Without
this change, the link info capabilities like the CRC flag will be
retained for the next hotplugged device. Now if, a MST is connected
after a SST connection, the base panel will have the CRC value of the
previous sink, which will be invalid for the case when the hub doesn't
support CRC.

Change-Id: I3a820070fd5006a707328b4d192893a465c04448
Signed-off-by: Soutrik Mukhopadhyay <quic_mukhopad@quicinc.com>