asoc: codecs: wsa883x: update INTR mask & clear reg's as volatile
Update Interrupt Mask and Clear registers as volatile
to prevent audio mute during pdr recovery.
As we observed these registers are been written during
regcache sync which resulting in stopping the swrm
interrupt. after updating these registers to volatile
swrm interrupts are received as expected when there is
any issue with PA on.

Change-Id: I80bdf1556b1e07ec4751da5eb4614d30ad11f5a0
Signed-off-by: Ganapathiraju Sarath Varma <quic_ganavarm@quicinc.com>